+++ src/openshift/cluster.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See LICENSE file in the project root for license information.
  *-----------------------------------------------------------------------------------------------*/
 
-import { window, commands, env, QuickPickItem, ExtensionContext, Terminal, Uri, workspace } from 'vscode';
+import { window, commands, env, QuickPickItem, ExtensionContext, Terminal, Uri, workspace, WebviewPanel } from 'vscode';
 import { Command } from '../odo/command';
 import OpenShiftItem, { clusterRequired } from './openshiftItem';
 import { CliExitData, CliChannel } from '../cli';
@@ -108,8 +108,11 @@ export class Cluster extends OpenShiftItem {
     }
 
     @vsCommand('openshift.explorer.addCluster')
-    static add(): void {
-        ClusterViewLoader.loadView('Add OpenShift Cluster');
+    static async add(value: string): Promise<void> {
+        const webViewPanel: WebviewPanel = await ClusterViewLoader.loadView('Add OpenShift Cluster');
+        if(value?.length > 0){
+            await webViewPanel.webview.postMessage({action: 'cluster', param: value});
+        }
     }
 
     @vsCommand('openshift.explorer.stopCluster')
@@ -119,8 +122,8 @@ export class Cluster extends OpenShiftItem {
         let crcBinary;
         const crcPath = workspace.getConfiguration('openshiftConnector').get('crcBinaryLocation');
         if(crcPath) {
-            newPathPrompt = { label: '$(plus) Provide different crc binary path'};
-            pathSelectionDialog = await window.showQuickPick([{label:`${crcPath}`, description: 'Fetched from settings'}, newPathPrompt], {placeHolder: 'Select CRC binary path', ignoreFocusOut: true});
+            newPathPrompt = { label: '$(plus) Provide different OpenShift Local file path'};
+            pathSelectionDialog = await window.showQuickPick([{label:`${crcPath}`, description: 'Fetched from settings'}, newPathPrompt], {placeHolder: 'Select OpenShift Local file path', ignoreFocusOut: true});
         }
         if(!pathSelectionDialog) return;
         if (pathSelectionDialog.label === newPathPrompt.label) {
@@ -129,14 +132,14 @@ export class Cluster extends OpenShiftItem {
                 canSelectFolders: false,
                 canSelectMany: false,
                 defaultUri: Uri.file(Platform.getUserHomePath()),
-                openLabel: 'Add crc binary path.',
+                openLabel: 'Add OpenShift Local file path.',
             });
             if (!crcBinaryLocation) return null;
             crcBinary = crcBinaryLocation[0].fsPath;
         } else {
             crcBinary = crcPath;
         }
-        const terminal: Terminal = WindowUtil.createTerminal('OpenShift: Stop CRC', undefined);
+        const terminal: Terminal = WindowUtil.createTerminal('OpenShift: Stop OpenShift Local', undefined);
             terminal.sendText(`${crcBinary} stop`);
             terminal.show();
         return;