+++ src/webview/devfile-registry/app/cardItem.tsx
@@ -121,6 +121,14 @@ export class CardItem extends React.Component<DevFileProps, {
     };
 
     copyClicked = (isClicked: boolean): void => {
+        if (isClicked) {
+            VSCodeMessage.postMessage(
+                {
+                    'action': 'telemeteryCopyEvent',
+                    'devFileName': this.props.compDescription.Devfile.metadata.name
+                }
+            )
+        }
         this.setState({
             copyClicked: isClicked
         });
@@ -137,7 +145,7 @@ export class CardItem extends React.Component<DevFileProps, {
                 </TextContent>
                 <Badge key={this.props.compDescription.Devfile.metadata.name + '-badge'}
                     className={clsx(this.props.cardItemStyle.badge, this.props.cardItemStyle.headerBadge)}
-                    overlap='rectangle'
+                    overlap='rectangular'
                     variant='standard'
                     showZero={false}>
                     {this.props.compDescription.Devfile.starterProjects.length}
@@ -347,7 +355,7 @@ export class CardItem extends React.Component<DevFileProps, {
                                     className={index === 0 ?
                                         clsx(this.props.cardItemStyle.badge, this.props.cardItemStyle.firstBadge)
                                         : this.props.cardItemStyle.badge}
-                                    overlap='rectangle'
+                                    overlap='rectangular'
                                     variant='standard'
                                 >
                                     {tag}